+++ Layout.mm
@@ -56,7 +56,7 @@
 }
 
 //-----------------------------------------------------------------------------
-- (void const *)imageAt:(uint32_t)location
+- (void const *)imageAt:(uint64_t)location
 {
   auto p = (uint8_t const *)[dataController.realData bytes];
   return p ? p + location : NULL;
